backend/requirements.txt
Django>=4.2.11,<5.0  # SECURITY: Updated to patch known vulnerabilities
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-ratelimit==4.1.0
python-dotenv==1.0.0
colorlog>=6.7.0
Pillow>=12.0.0
psycopg2-binary>=2.9.11
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
redis==5.0.1
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2
numpy>=1.24.0
requests>=2.31.0
gunicorn>=21.2.0

